AzureChatServer/requirements.txt
django
pyjwt
django-cors-headers
mysqlclient
openai
